requirements.txt
tensorflow>=2.13.0
keras>=2.13.1
numpy>=1.21.0
opencv-python>=4.8.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.3.0
Pillow>=9.0.0
pandas>=1.5.0
tqdm>=4.65.0